+++ src/main.tsx
@@ -4,13 +4,14 @@ import {
   cycleCurrentPaletteOffset,
   setPalette,
 } from "@/camera/palette";
+import BigNumber from "bignumber.js";
 import p5 from "p5";
 import React from "react";
 import ReactDOMClient from "react-dom/client";
 import { getIterationTimeAt } from "./aggregator/aggregator";
 import { setP5 } from "./canvas-reference";
 import { extractMandelbrotParams } from "./lib/params";
-import { calcVars, startCalculation } from "./mandelbrot";
+import { startCalculation } from "./mandelbrot";
 import {
   cycleMode,
   getCurrentParams,
@@ -37,6 +38,7 @@ import {
   readSettingsFromStorage,
 } from "./store/sync-storage/settings";
 import "./style.css";
+import type { MandelbrotParams } from "./types";
 import { AppRoot } from "./view/app-root";
 import { prepareWorkerPool } from "./worker-pool/pool-instance";
 import { getProgressData } from "./worker-pool/worker-pool";
@@ -114,6 +116,38 @@ const calcInteractiveZoomFactor = (p: p5) => {
 
   const minSize = 20;
   return Math.max(zoomFactor, minSize / p.width);
+};
+
+const calcVars = (
+  mouseX: number,
+  mouseY: number,
+  width: number,
+  height: number,
+  currentParams: MandelbrotParams = getCurrentParams(),
+) => {
+  // [-1, 1]に変換
+  const normalizedMouseX = new BigNumber(2 * mouseX).div(width).minus(1);
+  const normalizedMouseY = new BigNumber(2 * mouseY).div(height).minus(1);
+
+  const scaleX = width / Math.min(width, height);
+  const scaleY = height / Math.min(width, height);
+
+  const complexMouseX = currentParams.x.plus(
+    normalizedMouseX.times(currentParams.r).times(scaleX),
+  );
+  const complexMouseY = currentParams.y.minus(
+    normalizedMouseY.times(currentParams.r).times(scaleY),
+  );
+
+  const r = currentParams.r;
+  const N = currentParams.N;
+
+  return {
+    mouseX: complexMouseX,
+    mouseY: complexMouseY,
+    r,
+    N,
+  };
 };
 
 const sketch = (p: p5) => {
